% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHReal.R
\name{as.data.frame.mHReal}
\alias{as.data.frame.mHReal}
\title{Dataframe represetation of mHReal}
\usage{
\method{as.data.frame}{mHReal}(res)
}
\arguments{
\item{res}{S3-object of mHReal}
}
\description{
The realization of Hawkes model is represented by dataframe like ouput.
}
