% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhreal.R
\name{as.matrix.mhreal}
\alias{as.matrix.mhreal}
\title{Matrix represetation of mhreal}
\usage{
\method{as.matrix}{mhreal}(res)
}
\arguments{
\item{res}{S3-object of mhreal}
}
\description{
The realization of Hawkes model is represented by matrix like ouput.
}
