% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhreal.R
\name{dense_lambda}
\alias{dense_lambda}
\title{Get a lambda process with more dense time step}
\usage{
dense_lambda(arrival, lambda, beta, dt = NULL, ...)
}
\arguments{
\item{arrival}{a vector of arrival times.}

\item{lambda}{a vector of lambda processs. It should be a right continuous version.}

\item{beta}{a decaying parameter lambda.}

\item{dt}{a small step size on time horizon.}
}
\description{
Since the intenisty process is exponencially decaying, to fully describe the intensity processs,
we need more dense time line.
This function gives a lambda process with more dense time line.
}
\examples{
MU1 <- 0.3; ALPHA1 <- 1.5; BETA1 <- 2
mhspec1 <- new("mhspec", MU=MU1, ALPHA=ALPHA1, BETA=BETA1)
# Simulate with mhsim funciton.
res1 <- mhsim(mhspec1,  n=100)
dl <- dense_lambda(res1$arrival, res1$lambda, BETA1)
}
