% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHFit.R
\docType{methods}
\name{logLik,mHSpec-method}
\alias{logLik,mHSpec-method}
\title{Compute the loglikelihood function}
\usage{
\S4method{logLik}{mHSpec}(object, inter_arrival, jump_type = NULL,
  mark = NULL, LAMBDA0 = NULL)
}
\arguments{
\item{inter_arrival}{Inter-arrival times of events. Includes inter-arrival for events that occur in all dimensions. Start with zero.}

\item{jump_type}{a vector of dimensions. Distinguished by numbers, 1, 2, 3, and so on. Start with zero.}

\item{mark}{a vector of mark (jump) sizes. Start with zero.}

\item{LAMBDA0}{The starting values of lambda. Must have the same dimensional matrix (n by n) with \code{mHSpec}.}
}
\description{
This is a generic function.
The loglikelihood of the ground process of the Hawkes model.
(The estimation for jump distribution is not provided.)
}
\examples{
# construct a mHSpec
MU1 <- 0.2; ALPHA1 <- 1.0; BETA1 <- 2; ETA1 <- 0.2
JUMP1 <- function(n,...) rgeom(n, 0.7) + 1
mHSpec1 <- new("mHSpec", MU=MU1, ALPHA=ALPHA1, BETA=BETA1, ETA=ETA1, Jump =JUMP1)
# simualte a path
res1 <- mHSim(mHSpec1,  LAMBDA0 = MU1, n=1000)
inter_arrival <- res1$inter_arrival
mark <- res1$mark
# compute a loglikelihood function with parameter values in mHSpec1
# LAMBDA0 = MU1 is a naive way of starting point choice.
logLik(mHSpec1, LAMBDA0 = MU1, inter_arrival = inter_arrival, mark = mark)

}
\seealso{
\code{\link{mHSpec-class}}, \code{\link{mHFit,mHSpec-method}}
}
