% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhmoment.R
\docType{methods}
\name{mean_lambda,mhspec-method}
\alias{mean_lambda,mhspec-method}
\title{Compute the long-run mean of lambda}
\usage{
\S4method{mean_lambda}{mhspec}(object, mean_jump = NULL,
  sample_size = 10000, seed_value = 903)
}
\arguments{
\item{object}{mhspec, a one-dimensional or two-dimensional symmetricl model}

\item{mean_jump}{the mean of jump distribution. If not specified, simulated mean value will be used.}

\item{sample_size}{the number of simulation to compute the mean of jump and squared jump}

\item{seed_value}{seed for random number generation.}
}
\description{
This method only works for a one-dimensional or two-dimensional symmetric model.
}
