% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHSim.R
\docType{methods}
\name{mHSim,mHSpec-method}
\alias{mHSim,mHSpec-method}
\title{Simulate an n-dimensional marked Hawkes process}
\usage{
\S4method{mHSim}{mHSpec}(object, LAMBDA0 = NULL, n = 1000)
}
\arguments{
\item{object}{mHSpec}

\item{LAMBDA0}{The starting values of lambda. Must have the same dimensional matrix (n by n) with mHSpec.}

\item{n}{The number of observations.}
}
\description{
This is a generic function.
The mark (jump) structure may or may not be included.
It returns an object of class 'mHreal'.
}
\examples{
# example for one dimensional Hawkes process (without mark)
# Define the model.
MU1 <- 0.3; ALPHA1 <- 1.5; BETA1 <- 2
mHSpec1 <- new("mHSpec", MU=MU1, ALPHA=ALPHA1, BETA=BETA1)
# Simulate with mHSim funciton.
res1 <- mHSim(mHSpec1,  n=100)
summary(res1)
as.matrix(res1)

# example for two dimensional Hawkes process
# Define the model.
MU2 <- matrix(c(0.2), nrow = 2)
ALPHA2 <- matrix(c(0.75, 0.92, 0.92, 0.75), nrow = 2, byrow=TRUE)
BETA2 <- matrix(c(2.25, 2.25, 2.25, 2.25), nrow = 2, byrow=TRUE)
ETA2 <- matrix(c(0.19, 0.19, 0.19, 0.19), nrow = 2, byrow=TRUE)
JUMP2 <- function(n,...) rgeom(n, 0.65) + 1   # mark size follows a geometric distribution
mHSpec2 <- new("mHSpec", MU=MU2, ALPHA=ALPHA2, BETA=BETA2, ETA=ETA2, Jump =JUMP2)
# Simulate with mHSim function.
LAMBDA0 <- matrix(c(0.1, 0.1, 0.1, 0.1), nrow = 2, byrow=TRUE)
res2 <- mHSim(mHSpec2, LAMBDA0 = LAMBDA, n = 100)
summary(res2)
as.matrix(res2)
}
