% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHawkesSpec.R
\docType{class}
\name{mHSpec-class}
\alias{mHSpec-class}
\title{An S4 calss to represent the specification of a marked Hawkes model}
\description{
An S4 calss to represent the specification of a marked Hawkes model
}
