% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHVolatility.R
\docType{data}
\name{mHSpec2}
\alias{mHSpec2}
\title{Compute the variance of Hawkes difference process}
\format{An object of class \code{mHSpec} of length 1.}
\usage{
mHSpec2
}
\arguments{
\item{object}{mHSpec a two dimensional symmetricl model}

\item{time_length}{time horizon}

\item{mean_jump}{the mean of jump distribution. If not specified, simulated mean value will be used.}

\item{mean_jump_square}{the mean of the square of jump distribution. If not specified, simulated value will be used.}

\item{sample_size}{the number of simulation to compute the mean of jump and squared jump}

\item{seed_value}{seed for random number generation.}
}
\description{
This method computes Var(N1(t) - N2(t)).
This method only works for a two-dimensional symmetric model.
Assume that a two-dimensional Hawkes process describes the tick dynamics of financial price process.
One of the two Hawkes processes is responsible for the upward movement and the other for the downward movement.
Therefore, the difference between two process, N1 - N2, describes the tick price process.
This method is useful in quantitative finance and financial econometics.
}
\examples{
# two dimensional symmetric Hawkes model
MU2 <- matrix(c(0.2), nrow = 2)
ALPHA2 <- matrix(c(0.75, 0.92, 0.92, 0.75), nrow = 2, byrow=TRUE)
BETA2 <- matrix(c(2.90, 2.90, 2.90, 2.90), nrow = 2, byrow=TRUE)
ETA2 <- matrix(c(0.19, 0.19, 0.19, 0.19), nrow = 2, byrow=TRUE)
JUMP2 <- function(n,...) rgeom(n, 0.7) + 1
mHSpec2 <- new("mHSpec", MU=MU2, ALPHA=ALPHA2, BETA=BETA2, ETA=ETA2, Jump =JUMP2)
mHVar(mHSpec2, 1)
}
\keyword{datasets}
