% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHreal.R
\name{plotlambda}
\alias{plotlambda}
\title{Plot exponentially decaying lambda process}
\usage{
plotlambda(arrival, lambda, beta, dt = NULL, ...)
}
\arguments{
\item{arrival}{a vector of arrival times.}

\item{lambda}{a vector of lambda processs.}

\item{beta}{a decaying parameter lambda.}

\item{dt}{a small step size on time horizon.}
}
\description{
This plot method describes the exponentially decaying lambda (intensity) process.
}
\examples{

}
