% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhreal.R
\name{print.mhreal}
\alias{print.mhreal}
\title{Print function for mhreal}
\usage{
\method{print}{mhreal}(res, n = 20)
}
\arguments{
\item{res}{S3-object of mhreal}

\item{n}{number of rows to diplay}
}
\description{
Print the summary of the realization of the Haweks model.
}
