% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhreal.R
\name{summary.mhreal}
\alias{summary.mhreal}
\title{Summary function fo mhreal}
\usage{
\method{summary}{mhreal}(res, n = 20)
}
\arguments{
\item{res}{S3-object of mhreal}

\item{n}{number of rows to diplay}
}
\description{
This function presents the summary of the Hawkes realization.
}
