% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_spec.R
\name{valid_mhspec}
\alias{valid_mhspec}
\title{Check the validity of mhspec}
\usage{
valid_mhspec(object)
}
\arguments{
\item{object}{S4-class of mhspec}
}
\description{
This function checks the validity of mhspec.
For one dimensional case, if one of the parameter is not matrix, then all parameters should not be matrix.
If one of the parameter is atomic, then all parameters should be atomic.
If all parameters are matrix, then the nrow should be equal.
The number of columns of MU matrix should be one.
ALPHA, BETA, ETA matrices should be sqaure matrices.
The dimension of the model is less than 10.
}
