% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhreal.R
\name{as.data.frame.mhreal}
\alias{as.data.frame.mhreal}
\title{Dataframe represetation of mhreal}
\usage{
\method{as.data.frame}{mhreal}(res)
}
\arguments{
\item{res}{S3-object of mhreal}
}
\description{
The realization of Hawkes model is represented by dataframe like ouput.
}
