% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHReal.R
\name{as.matrix.mHReal}
\alias{as.matrix.mHReal}
\title{Matrix represetation of mHReal}
\usage{
\method{as.matrix}{mHReal}(res)
}
\arguments{
\item{res}{S3-object of mHReal}
}
\description{
The realization of Hawkes model is represented by matrix like ouput.
}
