% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHreal.R
\name{get_lc_lambda}
\alias{get_lc_lambda}
\title{Get left continuous version of lambda process}
\usage{
get_lc_lambda(res)
}
\arguments{
\item{res}{\code{mHreal} an S3 class contains the realized lambda processes.}
}
\description{
The realized version of the lambda process in \code{mHreal} is right continuous version.
If the left continuous version is needed, this function is applied.
}
