% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhreal.R
\name{get_lc_lambda}
\alias{get_lc_lambda}
\title{Get left continuous version of lambda process}
\usage{
get_lc_lambda(res)
}
\arguments{
\item{res}{\code{mhreal} an S3 class contains the realized lambda processes.}
}
\value{
The left continuous version of lambda components as a matrix.
}
\description{
The realized version of the lambda process in \code{mhreal} is right continuous version.
If the left continuous version is needed, this function is applied.
}
\examples{
# Define the model.
MU1 <- 0.3; ALPHA1 <- 1.5; BETA1 <- 2
mhspec1 <- new("mhspec", MU=MU1, ALPHA=ALPHA1, BETA=BETA1)
# Simulate with mhsim funciton.
res1 <- mhsim(mhspec1,  n=100)
get_lc_lambda(res1)

}
