% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHFit.R
\docType{methods}
\name{mHFit,mHSpec-method}
\alias{mHFit,mHSpec-method}
\title{Perform a maximum likelihood estimation}
\usage{
\S4method{mHFit}{mHSpec}(object, inter_arrival, jump_type = NULL,
  mark = NULL, LAMBDA0 = NULL, constraint = FALSE, method = "BFGS")
}
\arguments{
\item{inter_arrival}{Inter-arrival times of events. Includes inter-arrival for events that occur in all dimensions. Start with zero.}

\item{jump_type}{a vector of dimensions. Distinguished by numbers, 1, 2, 3, and so on. Start with zero.}

\item{mark}{a vector of mark (jump) sizes. Start with zero.}

\item{LAMBDA0}{The starting values of lambda. Must have the same dimensional matrix (n by n) with mHSpec.}

\item{constraint}{boolean. Set a constraint or not.}

\item{method}{method for optimization. For more information, see \code{\link[maxLik]{maxLik}}.}
}
\description{
This function uses \code{\link[maxLik]{maxLik}} for the optimizer.
}
\examples{
# Generate sample path
MU1 <- 0.3; ALPHA1 <- 1.5; BETA1 <- 2
mHSpec1 <- new("mHSpec", MU=MU1, ALPHA=ALPHA1, BETA=BETA1)
res1 <- mHSim(mHSpec1,  n=5000)

# Perform maximum likelihood estimation with a starting point defined by mHSpec0.
mHSpec0 <- new("mHSpec", MU=0.2, ALPHA=1.2, BETA=1.8)
mle <- mHFit(mHSpec0, inter_arrival = res1$inter_arrival)
summary(mle)
}
