% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhsim.R
\docType{methods}
\name{mhsim,mhspec-method}
\alias{mhsim,mhspec-method}
\title{Simulate an n-dimensional (marked) Hawkes process}
\usage{
\S4method{mhsim}{mhspec}(object, LAMBDA0 = NULL, n = 1000)
}
\arguments{
\item{object}{\code{\link{mhspec-class}}. This object includes the parameter values.}

\item{LAMBDA0}{the starting values of lambda (intensity process). Must have the same dimensional matrix (n by n) with the parameters in mhspec.}

\item{n}{the number of observations.}
}
\description{
The method simulate n-dimensional marked Hawkes processes.
The object \code{\link{mhspec-class}} contains the parameter values such as alpha, beta, eta.
The mark (jump) structure may or may not be included.
It returns an object of class 'mHreal'.
}
\examples{
# Example for one dimensional Hawkes process (without mark)
# Simple simulation for example
mhsim()
mhsim(dimens = 2)

# Define a one-dimensional model.
MU1 <- 0.3; ALPHA1 <- 1.5; BETA1 <- 2
mhspec1 <- new("mhspec", MU=MU1, ALPHA=ALPHA1, BETA=BETA1)
# Simulate with mhsim funciton.
res1 <- mhsim(mhspec1,  n=100)
summary(res1)
as.matrix(res1)

# Example for two dimensional Hawkes process
# Define a two-dimensional model.
MU2 <- matrix(c(0.2), nrow = 2)
ALPHA2 <- matrix(c(0.75, 0.92, 0.92, 0.75), nrow = 2, byrow=TRUE)
BETA2 <- matrix(c(2.25, 2.25, 2.25, 2.25), nrow = 2, byrow=TRUE)
ETA2 <- matrix(c(0.19, 0.19, 0.19, 0.19), nrow = 2, byrow=TRUE)
JUMP2 <- function(n,...) rgeom(n, 0.65) + 1   # mark size follows a geometric distribution
mhspec2 <- new("mhspec", MU=MU2, ALPHA=ALPHA2, BETA=BETA2, ETA=ETA2, Jump =JUMP2)
# Simulate with mhsim function.
LAMBDA0 <- matrix(c(0.1, 0.1, 0.1, 0.1), nrow = 2, byrow=TRUE)
res2 <- mhsim(mhspec2, LAMBDA0 = LAMBDA0, n = 100)
class(res2)
summary(res2)
as.matrix(res2)
}
