% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhspec.R
\docType{class}
\name{mhspec-class}
\alias{mhspec-class}
\title{An S4 class to represent a marked Hawkes model}
\description{
This class represents a specification of a marked Hawkes model with exponential kernel.
The intensity of the ground process is defined by:
\deqn{\lambda(t) = \mu + \int \alpha / \beta (1 + (k - 1) \eta) exp( -\beta (t-u)) d N(t)}
}
\section{Slots}{

\describe{
\item{\code{MU}}{numeric value or matrix. Shoud be a matrix for two or larger dimensional model.}

\item{\code{ALPHA}}{numeric value or matrix. Shoud be a matrix for two or larger dimensional model.}

\item{\code{BETA}}{numeric value or matrix. Shoud be a matrix for two or larger dimensional model.}

\item{\code{ETA}}{numeric value or matrix. Shoud be a matrix for two or larger dimensional model.}

\item{\code{Jump}}{a function that generate a random number with specific distribution.}
}}

\examples{
MU2 <- matrix(c(0.2), nrow = 2)
ALPHA2 <- matrix(c(0.75, 0.92, 0.92, 0.75), nrow = 2, byrow=TRUE)
BETA2 <- matrix(c(2.25, 2.25, 2.25, 2.25), nrow = 2, byrow=TRUE)
ETA2 <- matrix(c(0.19, 0.19, 0.19, 0.19), nrow = 2, byrow=TRUE)
JUMP2 <- function(n,...) rgeom(n, 0.65) + 1
mhspec2 <- new("mhspec", MU=MU2, ALPHA=ALPHA2, BETA=BETA2, ETA=ETA2, Jump =JUMP2)
}
