% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHReal.R
\name{print.mHReal}
\alias{print.mHReal}
\title{Print function for mHReal}
\usage{
\method{print}{mHReal}(res, n = 20)
}
\arguments{
\item{res}{S3-object of mHReal}

\item{n}{number of rows to diplay}
}
\description{
Print the summary of the realization of the Haweks model.
}
