% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHReal.R
\name{summary.mHReal}
\alias{summary.mHReal}
\title{Summary function fo mHReal}
\usage{
\method{summary}{mHReal}(res, n = 20)
}
\arguments{
\item{res}{S3-object of mHReal}

\item{n}{number of rows to diplay}
}
\description{
This function presents the summary of the Hawkes realization.
}
